A hacky path to texture-provided diffuse light.
I've kept textures in this project since they were first added in a
much earlier lesson, and I added 1-2 extras -- but the tutorials cut
the texture parts out around the start of the lighting work, and are
now adding them back in in a different shaped way -- which wouldn't
be much of a problem, except I added on top of the previous stuff to
allow multiple textures to be used, in a really hacky gross way.

So I've got the latest lesson working (diffuse light values using the
color sampled from a texture), but in that hacky way. We'll save that
for now, and next up I'll unwind some of my gross work to make it more
like the tutorial wants.