Implement directional shadows w/ a shadow map.
- the debug quad only renders when you're very close to it, otherwise
  it renders as the skybox
- the edges have been smoothed out a bit but still don't look great
- it's really tough to choose a spot to cast light from that can see
  the whole scene but has enough precision to map all of it
- 8192x8192 shadow map, yikes!

I think I should be finding a way to cast a directional light much
closer to the player, which can be much smaller but still fairly
high resolution (and with high precision), and then use some kind of
simpler thing farther out. Even as simple as just not rendering the
shadows.

Maybe I'll have the lookAt target be the player and the x/y/z of the
light be offset by the player's coords...?